debug

+++ attack/attack.py
@@ -20,5 +20,5 @@ def generate_attack(attack, eps, model, x, target, loss=nn.CrossEntropyLoss()):
             adversary = LinfPGDAttack(model, loss_fn=loss, eps=eps, nb_iter=50, eps_iter=0.03, rand_init=True,
                                       clip_min=0.,
                                       clip_max=1., targeted=False)
-        adv_image = adversary(x, target)
+        adv_image = adversary(x, target.float())
     return adv_image
